Add a --repair option to commit.
It's possible to construct a variable scope with consecutive
slashes (e.g. variable_scope('foo//')). When navigating through
a checkpoint with such a scope, there's no good way to navigate it
"correctly" so tf-explorer flattens out the structure in memory as
if there was just one slash but writes it out to disk with multiple
slashes as expected. This discrepancy is not good.

The --repair option finds all tensors with consecutive slashes and
renames them on disk to only have a single separating slash.